fixed dumb mistake

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: sholiak <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/16 10:15:50 by sholiak           #+#    #+#             */
-/*   Updated: 2019/05/12 17:02:42 by sholiak          ###   ########.fr       */
+/*   Updated: 2019/05/18 22:28:58 by sholiak          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
 	int len;
